+++ sdk/java/src/main/java/com/pulumi/mongodbatlas/outputs/AdvancedClusterAdvancedConfiguration.java
@@ -5,6 +5,7 @@ package com.pulumi.mongodbatlas.outputs;
 
 import com.pulumi.core.annotations.CustomType;
 import java.lang.Boolean;
+import java.lang.Double;
 import java.lang.Integer;
 import java.lang.String;
 import java.util.Objects;
@@ -14,7 +15,7 @@ import javax.annotation.Nullable;
 @CustomType
 public final class AdvancedClusterAdvancedConfiguration {
     /**
-     * @return The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively.`expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing.
+     * @return The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively. `expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing. This parameter is only supported for MongoDB version 6.0 and above.
      * 
      */
     private @Nullable Integer changeStreamOptionsPreAndPostImagesExpireAfterSeconds;
@@ -65,7 +66,7 @@ public final class AdvancedClusterAdvancedConfiguration {
      * * **Note**  A minimum oplog retention is required when seeking to change a cluster&#39;s class to Local NVMe SSD. To learn more and for latest guidance see [`oplogMinRetentionHours`](https://www.mongodb.com/docs/manual/core/replica-set-oplog/#std-label-replica-set-minimum-oplog-size)
      * 
      */
-    private @Nullable Integer oplogMinRetentionHours;
+    private @Nullable Double oplogMinRetentionHours;
     /**
      * @return The custom oplog size of the cluster. Without a value that indicates that the cluster uses the default oplog size calculated by Atlas.
      * 
@@ -89,7 +90,7 @@ public final class AdvancedClusterAdvancedConfiguration {
 
     private AdvancedClusterAdvancedConfiguration() {}
     /**
-     * @return The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively.`expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing.
+     * @return The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively. `expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing. This parameter is only supported for MongoDB version 6.0 and above.
      * 
      */
     public Optional<Integer> changeStreamOptionsPreAndPostImagesExpireAfterSeconds() {
@@ -154,7 +155,7 @@ public final class AdvancedClusterAdvancedConfiguration {
      * * **Note**  A minimum oplog retention is required when seeking to change a cluster&#39;s class to Local NVMe SSD. To learn more and for latest guidance see [`oplogMinRetentionHours`](https://www.mongodb.com/docs/manual/core/replica-set-oplog/#std-label-replica-set-minimum-oplog-size)
      * 
      */
-    public Optional<Integer> oplogMinRetentionHours() {
+    public Optional<Double> oplogMinRetentionHours() {
         return Optional.ofNullable(this.oplogMinRetentionHours);
     }
     /**
@@ -202,7 +203,7 @@ public final class AdvancedClusterAdvancedConfiguration {
         private @Nullable Boolean javascriptEnabled;
         private @Nullable String minimumEnabledTlsProtocol;
         private @Nullable Boolean noTableScan;
-        private @Nullable Integer oplogMinRetentionHours;
+        private @Nullable Double oplogMinRetentionHours;
         private @Nullable Integer oplogSizeMb;
         private @Nullable Integer sampleRefreshIntervalBiConnector;
         private @Nullable Integer sampleSizeBiConnector;
@@ -267,7 +268,7 @@ public final class AdvancedClusterAdvancedConfiguration {
             return this;
         }
         @CustomType.Setter
-        public Builder oplogMinRetentionHours(@Nullable Integer oplogMinRetentionHours) {
+        public Builder oplogMinRetentionHours(@Nullable Double oplogMinRetentionHours) {
 
             this.oplogMinRetentionHours = oplogMinRetentionHours;
             return this;
